backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
Pillow
aiohttp
scikit-learn
scipy
transformers
torch
openai
python-dotenv
langchain
langchain-openai
langchain-core
langgraph
pydantic
beautifulsoup4
lxml
roboflow
